# cmd/api/docs/swagger.yaml
definitions:
...
  helperStruct.Category:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  helperStruct.Product:
    properties:
      brand:
        type: string
      categoryid:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - brand
    - categoryid
    - description
    - name
    type: object
  helperStruct.ProductItem:
    properties:
      battery:
        type: integer
      camera:
        type: integer
      colour:
        type: string
      image:
        type: string
      price:
        type: integer
      productid:
        type: integer
      quantity:
        type: integer
      ram:
        type: integer
      screensize:
        type: number
      sku:
        type: string
      storage:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/category/add:
    post:
      consumes:
      - application/json
      description: Admin can create new category from admin panel
      operationId: create-category
      parameters:
      - description: New category name
        in: body
        name: category_name
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create new product category
      tags:
      - Product Category
  /admin/category/find/{id}:
    get:
      consumes:
      - application/json
      description: Users and admins can fetch details of a specific category using
        id
      operationId: find-category-by-id
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Fetch details of a specific category using category id
      tags:
      - Product Category
  /admin/category/listall:
    get:
      consumes:
      - application/json
      description: Admin, users and unregistered users can see all the available categories
      operationId: view-all-categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: View all available categories
      tags:
      - Product Category
  /admin/category/update/{id}:
    patch:
      consumes:
      - application/json
      description: Admin can update category details
      operationId: update-category
      parameters:
      - description: ID of the Category to be updated
        in: path
        name: id
        required: true
        type: string
      - description: category info
        in: body
        name: category_details
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can update category details
      tags:
      - Product Category
...
  /admin/product-item/add/:
    post:
      consumes:
      - application/json
      description: This endpoint allows an admin user to create a new product item.
      operationId: create-product-item
      parameters:
      - description: Product item details
        in: body
        name: product_item
        required: true
        schema:
          $ref: '#/definitions/helperStruct.ProductItem'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added new product item
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to add new product item
          schema:
            $ref: '#/definitions/response.Response'
      summary: Creates a new product item
      tags:
      - Product Item
  /admin/product-item/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a product item from the system
      operationId: delete-product-item
      parameters:
      - description: ID of the product item to be deleted
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Deletes a product item from the system
      tags:
      - Product Item
  /admin/product-item/show/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a product item by its ID
      operationId: find-product-item-by-id
      parameters:
      - description: Product item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieve a product item by ID
      tags:
      - Product Item
  /admin/product-item/update/{id}:
    patch:
      consumes:
      - application/json
      description: Update an existing product item with new information.
      operationId: update-product-item
      parameters:
      - description: ID of the productitem to be updated
        in: path
        name: Id
        required: true
        type: string
      - description: Product item information to update
        in: body
        name: product_item
        required: true
        schema:
          $ref: '#/definitions/helperStruct.ProductItem'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update a product item
      tags:
      - Product Item
  /admin/product/add:
    post:
      consumes:
      - application/json
      description: Admins can create new product listings
      operationId: create-product
      parameters:
      - description: new product details
        in: body
        name: new_product_details
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can create new product listings
      tags:
      - Product
  /admin/product/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This endpoint allows an admin  to delete a product by ID.
      operationId: delete-product
      parameters:
      - description: Product ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted product
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.Response'
      summary: Deletes a product by ID
      tags:
      - Product
  /admin/product/update/{id}:
    put:
      consumes:
      - application/json
      description: This endpoint allows an admin user to update a product's details.
      operationId: update-product
      parameters:
      - description: ID of the product to be updated
        in: path
        name: Id
        required: true
        type: string
      - description: Updated product details
        in: body
        name: updated_product_details
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Product'
      produces:
      - application/json
      responses:
        "202":
          description: Successfully updated product
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Unable to update product
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can update product details
      tags:
      - Product
...
  /listallproduct/:
    get:
      consumes:
      - application/json
      description: Admins and users can ses all available products
      operationId: user-view-all-products
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      - description: Search query string
        in: query
        name: query
        type: string
      - description: Filter criteria for the product items
        in: query
        name: filter
        type: string
      - description: Sorting criteria for the product items
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admins and users can see all available products
      tags:
      - Product
...
  /product-item/uploadimage/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Upload one or more images for a given product ID
      operationId: upload-image
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product image(s)
        in: formData
        name: images
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Handles the uploading of images for a given product ID
      tags:
      - Product Item
...
  /userdisaplayallproductItems/:
    get:
      consumes:
      - application/json
      description: view all product items for user
      operationId: user-view-all-product-items
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      - description: Search query string
        in: query
        name: query
        type: string
      - description: Filter criteria for the product items
        in: query
        name: filter
        type: string
      - description: Sorting criteria for the product items
        in: query
        name: sort_by
        type: string
      - description: Sorting in descending order
        in: query
        name: sort_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Handler function to view all product items
      tags:
      - Product Item
...
swagger: "2.0"